outfit_generator/outfit_generator/requirements.txt
# ML модели
transformers>=4.52.4,<5.0.0
torch>=2.8.0
torchvision>=0.23.0
openai-clip>=1.0.1
faiss-cpu>=1.7.4
mediapipe>=0.10.0
pillow>=10.0.0

# Web и API
fastapi>=0.109.0
uvicorn>=0.27.0
requests>=2.31.0
beautifulsoup4>=4.12.0
aiohttp>=3.9.0

# Обработка данных
pandas>=2.0.0
numpy<2.0
scikit-learn>=1.3.0

# Утилиты
python-dotenv>=1.0.0
pydantic>=2.5.0
loguru>=0.7.0

# Дополнительно (без tensorflow!)
accelerate>=0.25.0
sentencepiece>=0.1.99
protobuf>=3.20.0
